Updated copy on page, added links
I didn't do anything with the image links for panelist photos, and there
are book titles that need to be italicized under both instances of Lara
Hogan and Scott Jehl. I didn't remember how to do it.

I also deleted the text on line 23 for event-promo, as I think this is
where the graphic was meant to go.