Update index.html
Added an h2 (applied to Scale Your Agency) to the small sponsor block, but not sure if it will style in the same way as the stuff in the logistics blocks above it, or if there is an alternate styling you'd go with to help the phrase stand out. There is an h3 applied to Scale Your Agency in the larger sponsor block. What will that look like? 

Added wording about A List Apart: On Air to the page. 

Do we want to link the Pantheon logo in the small sponsor block to the large one on the page, or link out?